requirements.txt

fastapi==0.115.0
uvicorn==0.30.0
pandas==2.2.0
openpyxl==3.1.2
sqlalchemy==2.0.25
python-dotenv==1.0.0
httpx==0.27.0
openai==1.40.0
pydantic==2.7.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.0